Add `willSave` and `willSaveWaitUntil` capabilities
Note that these are not tested at all and we might end up removing if
it turns out not useful. I wrote these while reading the spec
and thinking what's necessary to support them. I originally thought
`TextEdit[]` from `willSaveWaitUntil` request is used to implement
format on save, but it doesn't seem like it because `rust-analyzer`
doesn't support this capability.
In fact, I don't have a server with this capability to test against.